added userdata as props for navbar
